requirements.txt
Django>=2.2
easy-thumbnails==2.8
click==8.0.2
easy-thumbnails==2.8
importlib-metadata==4.8.3
Pillow==8.4.0
filelock==3.4.1
flake8==2.3.0
keyring==23.4.1
platformdirs==2.4.0
rfc3986==1.5.0
zipp==3.6.0
